Add support for together.ai inference and implement tool calls using it (WIP)
In theory, specialized function calling models should be great. In practice,
this may be so for structured inputs. For editing free-form text for e.g.
WolframAlpha, at least NexusRaven is terrible.

Another problem that neither GPT or Claude models had is that for a typical
chat bot use case, the functions should not be ALWAYS called. This seems to be
surprisingly difficult for NexusRaven.

Next steps: Use Mistral derivatives for deciding when to call tools?
Then (depending on tool?), use NexusRaven to figure out the arguments.
Or just use Mistral for everything, with e.g. Anyscale's new JSON mode?