Try JSON format and multi-step approach for tool calling
It works! Sort of, and unfortunately still not nearly as well as even
Claude instant (with XML) or GPT-3.5 (native). Perhaps I'll revisit this again,
since I can't be the only one who is trying to run function calling on
top of them...

I also added Anyscale's Mixtral model with function calling. It doesn't seem to
be much better, so maybe the Mixtral models are just not that well-suited for it?