Prepare for generative search
The goal is to ingest English Wikipedia, or failing that, parts of it.
Then, let's use vector search (with HYDE?) to provide LLMs relevant facts.
This should (in theory) reduce the risk of hallucinations.

Using cloud embeddings endpoints for entire Wikipedia is too expensive.
Instead, I built a prototype system to rent GPUs from Runpod for cheap.
I also created a Docker image that can run best open embedding model and
provide a REST API for it.

Not shown here, I've also ingested the Wikipedia into a normal database
(mongodb, with dumpster-dive and wtf_wikipedia). Automated pipeline for that
would be nice, but we'll see.

Next up: Ingesting Wikipedia from Mongodb and embedding it.
Storing embeddings in... Qdrant? Weaviate?

And hopefully some cleanup, this is a mess.